+++ crates/forge/src/scarb.rs
@@ -1,16 +1,10 @@
 use crate::scarb::config::ForgeConfigFromScarb;
 use anyhow::{Context, Result, anyhow};
-use cairo_lang_sierra::program::VersionedProgram;
-use camino::Utf8Path;
 use configuration::Config;
 use configuration::core::Profile;
-use forge_runner::package_tests::TestTargetLocation;
-use forge_runner::package_tests::raw::TestTargetRaw;
-use scarb_api::{ScarbCommand, test_targets_by_name};
-use scarb_metadata::{Metadata, PackageId, PackageMetadata};
+use scarb_api::{ScarbCommand};
+use scarb_metadata::{Metadata, PackageId};
 use scarb_ui::args::{FeaturesSpec, PackagesFilter, ProfileSpec};
-use std::fs;
-use std::io::ErrorKind;
 
 pub mod config;
 
@@ -90,48 +84,4 @@ fn build_test_artifacts_with_scarb(
         .run()
         .context("Failed to build test artifacts with Scarb")?;
     Ok(())
-}
-
-#[tracing::instrument(skip_all, level = "debug")]
-pub fn load_test_artifacts(
-    target_dir: &Utf8Path,
-    package: &PackageMetadata,
-) -> Result<Vec<TestTargetRaw>> {
-    let mut targets = vec![];
-
-    let dedup_targets = test_targets_by_name(package);
-
-    for (target_name, target) in dedup_targets {
-        let tests_location =
-            if target.params.get("test-type").and_then(|v| v.as_str()) == Some("unit") {
-                TestTargetLocation::Lib
-            } else {
-                TestTargetLocation::Tests
-            };
-
-        let target_file = format!("{target_name}.test.sierra.json");
-        let sierra_program_path = target_dir.join(target_file);
-
-        match fs::read_to_string(&sierra_program_path) {
-            Ok(value) => {
-                let versioned_program = serde_json::from_str::<VersionedProgram>(&value)?;
-
-                let sierra_program = match versioned_program {
-                    VersionedProgram::V1 { program, .. } => program,
-                };
-
-                let test_target = TestTargetRaw {
-                    sierra_program,
-                    sierra_program_path,
-                    tests_location,
-                };
-
-                targets.push(test_target);
-            }
-            Err(err) if err.kind() == ErrorKind::NotFound => {}
-            Err(err) => Err(err)?,
-        }
-    }
-
-    Ok(targets)
 }
